requirements.txt
uvicorn==0.30.1
fastapi==0.110.3
python-dotenv==1.0.0
crewai==0.51.1

sqlalchemy==2.0.31
tiktoken==0.7.0

durationpy==0.6
async-timeout
psycopg2-binary==2.9.9
anthropic
pandas
langchain-core
langchain-community
langchain-aws
langchain-anthropic
langgraph
boto3